feat: update version
