+++ packages/rax-recyclerview/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rax-recyclerview",
-  "version": "1.3.4",
+  "version": "1.3.5",
   "description": "RecyclerView component for Rax.",
   "license": "BSD-3-Clause",
   "main": "lib/index.js",
@@ -61,7 +61,7 @@
     "prop-types": "^15.7.2",
     "rax-refreshcontrol": "^1.0.0",
     "rax-scrollview": "^3.3.3",
-    "rax-view": "^1.0.2",
+    "rax-view": "^2.0.0",
     "universal-env": "^3.0.0"
   },
   "peerDependencies": {
@@ -85,7 +85,7 @@
     "rax-slider": "^1.1.0-0",
     "rax-tab-panel": "^1.0.1",
     "rax-test-renderer": "^1.0.0",
-    "rax-text": "^1.0.0",
+    "rax-text": "^2.0.0",
     "universal-transition": "^1.0.0",
     "typescript": "^3.5.3"
   }
